Move the existing tab management functionality to a separate class
